Manage Audit GitHub Workflow for Rust repositories with terraform
The `audit.yaml` workflow in Rust repositories has been toilsome to
update when new `cargo-deny` releases are made. To work around this,
configuration for the `cargo-deny` version was centrally managed via an
organization secret. This secret was managed with terraform as of #54.

Secrets are not visible to PRs from forks, which meant that PRs to
artichoke/artichoke would fail the audit workflow. To work around this,
manual PRs were issued across all Rust repositories to add a fallback
`cargo-deny` version in the bash script for the installer task. See
artichoke/artichoke#1013 for an example PR.

The 0.8.9 release of `cargo-deny` fixes a bug caused by the main branch
changing of a runtime dependency on the rust advisory DB. This has the
potential to break the workflow when the fallback version is invoked.

To fix this issue moving forward, this commit manages the
`.github/workflows/audit.yaml` file for all Rust repositories in the
artichoke organization. The `cargo-deny` version is controlled with a
terraform local and there is no dependency on any organization secrets.

When the `cargo-deny` version changes, operators should toggle the
`cargo_deny_force_bump` local to true and update the `cargo-deny`
version in the `cargo_deny_version` local.

A subsequent `terraform apply` will create new branches in all Rust
repos with an updated GitHub Actions workflow file. These branches can
be turned into PRs which will run status checks and update the workflow.

This commit updates the `cargo-deny` version to 0.8.9 and removes the
`CARGO_DENY_VERSION` organization secret.