Rewrite AWS module for granting read-only S3 access to GitHub Actions
Rename the `repository-assume-role` module to
`github-actions-s3-bucket-read-only-access` to limit the scope of the
module to what it actually does.

Follow best practices in upstream documentation for developing modules:

https://www.terraform.io/language/modules/develop/structure

- `main.tf`, `outputs.tf`, `variables.tf`, `README.md`.
- Add `description` to all variables and outputs.
- Parameterize the module on GitHub org, repo, and S3 bucket.
- Use `name_prefix` instead of `name` for IAM roles and policies.

Some other quality improvements:

- Remove `versions.tf` in module definition.
- Move all variables to `variables.tf` in `aws` workspace.
- Move all outputs to `outputs.tf` in `aws` workspace.
- Use `artichoke-forge` environment as a default tag for the `aws`
  provider.

This commit resulted in the entire IAM setup being replaced. Reflect the
new role to assume in GitHub Actions workflows.