Cease using `s3:kms` with CMK as default encryption for S3 buckets
Remove KMS keys from the following modules:

- `access-logs-s3-bucket`
- `private-s3-bucket`
- `private-archive-s3-bucket`

The Artichoke organization already has an SCP to prevent unencrypted
uploads from making it into S3 buckets. SSE with the S3-managed key is
sufficient and easier to use (simply pass `--sse AES256` on the `asw s3`
CLI) and does not require passing around KMS key IDs.

This removes a bunch of complexity and does not sacrifice security in
any meaningful way for this project.

This change switches the default bucket encryption configuration to
AES256 with the SSE-S3 key management strategy. The previously
provisioned KMS keys are destroyed.

This has the consequence of rendering some bucket objects unreachable.
Some of these are access logs which will age out. Others are terraform
state versions, which are lost.