Add custom, more permissive KMS key polict in `kms-key` module
Add a non-default KMS key policy in the `kms-key` module. This key
policy contains the same stanzas as the default key policy but also
allows specifiying a list of roles that should always have acceess to
the KMS key for encryption and decryption.

This list of additional roles is by default empty which makes these
changes a no-op for existing users of this module.