Fix IAM Policy for GitHub repository backup role
This polict was not granting any permissions because `s3:prefix` is not
available during calls which require `s3:PutObject`.

This commit changes the IAM policy to embed the path prefix in the ARN
of the resource target. Resources that get granted now look like:

    aws:arn:...:artichoke-forge-github-backups-us-west-2/project-infrastructure/*

This resource has the intended effect of limiting writes to a prefix
path in the bucket.