Add `s3-backup.yaml` GitHub repo orchestration with terraform
Add a new instance of the `update-github-repository-file` module to
place a GitHub Actions workflow called `s3-backup.yaml` into Artichoke
repositories.

This code creates a workflow that will generate a backup of the
repository, issues, comments, pulls, and releases and upload them to S3
as tarballs and zips. Releases are uploaded directly and not archived
and compressed.

The workflow is templated with the AWS IAM role and S3 bucket created in
the `aws` environment for backups. Backups are generated on a cron once
a month. The day for the cron is randomly chosen between 1 and 25 to
give some skew and make sure we don't hit GitHub rate limits.

The workflow uses the `github-backup` Python package to execute the
backup using a GitHub access token generated by the GitHub Actions
workflow. https://github.com/josegonzalez/python-github-backup

This backup code is inspired by this article, but using GitHub Actions
for cron and access keys instead:

- https://www.marcolancini.it/2021/blog-github-backups-with-ecs/

The workflow uses its workflow permissions to also acquire a set of AWS
credentials using OpenID Connect federation to make the uploads to S3.
See:

- #229
- #230

The resulting backups are uploaded to a repository-specific prefix and
are laid out like this:

    s3://artichoke-forge-github-backups-us-west-2/project-infrastructure/archives/tarballs/project-infrastructure-2022-02-20.tar.gz
    s3://artichoke-forge-github-backups-us-west-2/project-infrastructure/archives/zips/project-infrastructure-2022-02-20.zip
    s3://artichoke-forge-github-backups-us-west-2/project-infrastructure/releases/<release name 1>
    s3://artichoke-forge-github-backups-us-west-2/project-infrastructure/releases/<release name 2>

Backups are ecrypted with AWES256 S3-SSE and S3-managed keys.

This bit of terraform code has been tested and deployed in:

- #232
- https://github.com/artichoke/project-infrastructure/pull/232#issuecomment-1047019970