Update rustdoc workflow to deploy docs with GitHub Actions
Deploy generated Rust API documentation to GitHub Pages using the new
GitHub Actions deployment model instead of the `gh-pages` branch
deployment model:

- https://github.blog/2022-08-10-github-pages-now-uses-actions-by-default/
- https://github.blog/changelog/2022-07-27-github-pages-custom-github-actions-workflows-beta/

This approach follows the model in the `hugo` starter workflow:

- https://github.com/actions/starter-workflows/blob/edae9e95bf1c84aff6e1d34afcb7bde9288d0e25/pages/hugo.yml

Most updates in the workflow are to model it after the starter workflow
template. There are additional changes to Rust toolchain setup:

- Enforce nightly toolchain with the `RUSTUP_TOOLCHAIN` env variable
  which has high priority per https://rust-lang.github.io/rustup/overrides.html.
- Remove `rustup default nightly` command.
- No longer remove `rust-toolchain` file since `RUSTUP_TOOLCHAIN` has
  higher priority.
- Print out rustdoc verbose version with `rustdoc -Vv`.

Because generated docs are no longer stored in git, the workaround to
remove docs for large generated Rust sources is no longer required and
is removed.

`peaceiris/actions-gh-pages` is no longer used to deploy the docs.